[Bronze II] Title: 숫자, Time: 240 ms, Memory: 24020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -11,17 +11,18 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
-        int max = Math.max(a,b);
-        int min = Math.min(a,b);
+        long max = Math.max(a,b);
+        long min = Math.min(a,b);
         System.out.println(max==min? 0 : max-min - 1);
         StringBuilder sb = new StringBuilder();
-        for (int i = min+1; i < max; i++) {
+        for (long i = min+1; i < max; i++) {
             sb.append(i);
             if(i != max-1) sb.append(" ");
         }
         System.out.println(sb.toString());
     }
 }
+
